entries/omni-data/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omni-data-market-engine"
version = "0.1.0"
description = "The All-Seeing Eye — exhaustive, model-agnostic, script-only financial data ingestion and normalization engine for Open Athanor."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    {name = "Alembic AI"},
]
keywords = ["finance", "market-data", "ccxt", "options", "derivatives", "orderflow", "analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Topic :: Office/Business :: Financial :: Investment",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "ccxt>=4.0.0",
    "pydantic>=2.0.0",
    "pyzmq>=25.0.0",
    "aiohttp>=3.9.0",
    "mplfinance>=0.12.0b0",
    "plotly>=5.0.0",
    "kaleido>=0.2.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21.0",
]
redis = [
    "redis>=5.0.0",
]
all = [
    "omni-data-market-engine[redis]",
]

[project.scripts]
omni-data = "omni_data.orchestrator:cli"

[project.urls]
Homepage = "https://github.com/alembic-ai/opn_athnr_beta"
Repository = "https://github.com/alembic-ai/opn_athnr_beta"

[tool.setuptools.packages.find]
where = ["."]
include = ["omni_data*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
